feat(jest): 97% coverage for store/friends/mutations
